Improved Connectivity check state
