+++ lib/state_management/providers/class_providers.dart
@@ -4,13 +4,14 @@ import 'package:anonaddy/services/alias/alias_service.dart';
 import 'package:anonaddy/services/app_version/app_version_service.dart';
 import 'package:anonaddy/services/biometric_auth/biometric_auth_service.dart';
 import 'package:anonaddy/services/changelog_service/changelog_service.dart';
-import 'package:anonaddy/services/connectivity/connectivity_service.dart';
 import 'package:anonaddy/services/data_storage/offline_data_storage.dart';
 import 'package:anonaddy/services/domain/domains_service.dart';
 import 'package:anonaddy/services/domain_options/domain_options_service.dart';
 import 'package:anonaddy/services/recipient/recipient_service.dart';
 import 'package:anonaddy/services/username/username_service.dart';
+import 'package:anonaddy/state_management/connectivity/connectivity_state.dart';
 import 'package:anonaddy/state_management/domain_state_manager.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../alias_state_manager.dart';
@@ -29,7 +30,7 @@ final accessTokenServiceProvider = Provider((ref) => AccessTokenService());
 final recipientServiceProvider = Provider((ref) => RecipientService());
 final domainServiceProvider = Provider((ref) => DomainsService());
 final biometricAuthServiceProvider = Provider((ref) => BiometricAuthService());
-final connectivityServiceProvider = Provider((ref) => ConnectivityService());
+final connectivityState = Provider((ref) => ConnectivityState(Connectivity()));
 final offlineDataProvider = Provider((ref) => OfflineData());
 final changelogServiceProvider = Provider((ref) => ChangelogService());
 final appVersionServiceProvider = Provider((ref) => AppVersionService());
